renamed frontend and backend folders, added a database folder and created a custom database docker image, finalised the docker-compose.yml file, added license and code of conduct
