+++ README.md
@@ -1,20 +1,42 @@
-Om de backend op te starten moeten we eerst naar de backend folder gaan cd "Resolum api" en om alle node modules te installeren runnen we "npm install" en om de service op te starten runnen we "npm start" en als we dan naar localhost:3000 gaan runned hier onze backend en vergeet zeker niet de 'IP' en de 'PORT' variabele aan te maken in de .env folder waardat IP = resolum ip en PORT is 3000.
 
-Voor de frontend op te starten moeten we eerst naar de frontend folder gaan cd resolum_frontend en om alle node modules te installeren runnen we "npm install" en om de service op te starten runnen we "npm start" de terminal gaat dan vragen of je de service op een andere poort wilt runnen we confirmeren dit door "y" in de terminal te typen (react runned standaard op poort 3000 maar onze backend runned hier al op) automatisch gaat deze een venster openen, moest dit niet het geval zijn kan je altijd naar localhost:3001 surfen.
+## Installation
 
-Voor de database heb ik geen gebruik gemaakt van docker maar als je lokaal een database opstart met de volgende stukken code.
-create schema resolum;
-CREATE TABLE `resolum`.`clips2` (
-  `id` INT NOT NULL AUTO_INCREMENT,
-  `clip_path` VARCHAR(255) NULL,
-  `current_active` VARCHAR(255) NULL,
-  PRIMARY KEY (`id`));
-vergeet dan ook niet in de .env file de volgende variabelen aan te maken
-DB_HOST = localhost
-DB_USER = 'uw username'
-DB_PASSWORD = 'uw password'
-DB_SCHEME = resolum 
-Als de applicatie gerunned word zonder gebruikt te maken van de database moet lijn 28 in de file Resolum api/app.js in commentaar gezet worden
-'layers.insertClipInfo();' want deze lijn zorgt voor de insert query in de database.
+To start the backend service you first have to make sure you are in the backend folder 'backend'. To install all the node modules you have to run `npm install` (this can take a while). When all the modules are installed we can simply start the service by running `npm start`, the backend API will now run on `localhost:5000` (or if you have specified a different port in the .env file you have to navigate to that port). To make sure the backend is working optimally you have to run the Resolume REST API and fill in the ip address of your computer in the .env file of the backend service under the key IP.
 
-Om deze applicatie juist te gebruiken moet ook zeker de Resolum api aanstaan zodat de backend de juiste calls kan maken om da data binnen te halen.
+
+Starting the frontend is very simillar to the backend, first make sure you are in the right directory "frontend", then to install the node modules run `npm install` and to start the service run `npm start`.
+You can now access the frontend on `localhost:3000`.
+  
+
+## (Docker) Configuration
+
+To make sure the backend can communicate with the Resolume Webserver and the REST API we will have to fill in the IP address of the host computer in the 'docker-compose.yml' file. The backend is running a bridge network allowing the service to communicate with the host computer (this implies that the Resolume API has to run on the host computer).
+
+```yml
+environment:
+	- PORT=5000
+	- DB_HOST=db
+	- DB_USER=root
+	- DB_PASSWORD=root	
+	- DB_SCHEME=resolume
+	- IP=<IP ADDRESS>
+```
+In the docker-compose.yml file you will find the above structure in the 'backend' service. Here you have to replace <IP ADDRESS> with the IP address of the host computer.
+
+When you are done configuring the 'docker-compose.yml' file all you have to do is run `docker-compose up` in the directory of the 'docker-compose.yml' file and wait untill all the services are set up.
+
+## Usage
+
+To use the application you have to go to `localhost:3000`, here you will find the frontend. You can interact with the different clips and the changes will be reflected in your own Resolume arena.
+## Dependencies
+To run this entire application you will have to install [Docker](https://docs.docker.com/engine/install/).
+
+## License
+this project is released under the [MIT License](./LICENSE).
+
+## Contact
+Yassin Bendella
+If you have any questions, please feel free to send me a messsage on discord, linkedin or just through my email.
+YassinDella#7708
+[My Linkedin]()
+[yassin.bendella@student.ehb.be](yassin.bendella@student.ehb.be)
