+++ backend/.env
@@ -1,6 +1,6 @@
-PORT = 3000
+PORT = 5000
 DB_HOST = localhost
 DB_USER = root
 DB_PASSWORD = yassin
-DB_SCHEME = resolum 
-IP = 192.168.0.136
+DB_SCHEME = resolume 
+IP = 192.168.129.33
