+++ backend/Dockerfile
@@ -1,0 +1,13 @@
+FROM node:latest
+
+WORKDIR user/app
+
+COPY package*.json ./
+
+RUN npm install
+
+COPY ./ ./
+
+EXPOSE 5000
+
+CMD ["npm", "start"]
