+++ database/Dockerfile
@@ -1,0 +1,3 @@
+FROM mysql:5.7
+# https://stackoverflow.com/questions/44239766/dockerize-mysql-with-database-and-tables
+COPY ./init.sql /docker-entrypoint-initdb.d/
