+++ docker-compose.yml
@@ -1,0 +1,38 @@
+version: "3"
+
+services:
+
+  backend:
+    build: ./backend
+    restart: always
+    depends_on:
+      - db
+    ports:
+      - 5000:5000
+    environment:
+      - PORT=5000
+      - DB_HOST=db
+      - DB_USER=root
+      - DB_PASSWORD=root
+      - DB_SCHEME=resolume
+      - IP=192.168.129.33
+    networks:
+      - mynetwork
+
+  frontend:
+    build: ./frontend
+    ports: 
+      - 3000:3000
+    environment:
+      - REACT_APP_HOST=localhost:5000
+
+  db:
+    build: ./database
+    restart: always
+    environment:
+      MYSQL_ROOT_PASSWORD: root
+      MYSQL_DATABASE: resolume
+  
+networks:
+  mynetwork:
+    driver: bridge
