+++ frontend/.env
@@ -1,0 +1,1 @@
+REACT_APP_HOST=localhost:5000
