+++ frontend/src/components/layerComponent.js
