+++ frontend/src/index.js
@@ -4,6 +4,10 @@ import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 
+const host = process.env.REACT_APP_HOST
+console.log("backend host:")
+console.log(host)
+
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
