--- resolum_frontend/src/services/layerService.js
@@ -1,12 +1,0 @@
-import axios from'axios';
-
-const getLayerData = (id) => axios.get(`http://localhost:3000/layerdata/${id}`)
-const getClipData = () => axios.get(`http://localhost:3000/clipdata`)
-const setSelectedClip = (layerId,clipId) => axios.post(`http://localhost:3000/setselectedclip/${layerId}/${clipId}`)
-
-const layerService = {
-    getLayerData,
-    getClipData,
-    setSelectedClip
-}
-export default layerService;
